primera subida itscooter

+++ gap_itscooter/models/inherit_product.py
@@ -5,7 +5,7 @@ logger = logging.getLogger(__name__)
 class Products(models.Model):
     _inherit = "product.template"
 
-    Qty = fields.Integer(string="Cantidad total")
+    Qty = fields.Float(string="Cantidad total")
     Price_cost = fields.Float(string="Price")
     SKU = fields.Char(String='SKU')
     EAN = fields.Char(String='EAN')
